requirements.txt
APScheduler==3.10.4
python-dotenv==1.0.1
python-telegram-bot==21.6
python_kis==1.0.6
pytz==2022.1
fastapi==0.109.0
uvicorn==0.24.0
pytest==7.4.4
httpx==0.27.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
